requirements.txt
fastapi
uvicorn
websockets
winrt-Windows.Devices.Sensors
winrt-Windows.Devices.Enumeration
winrt-Windows.Foundation
winrt-Windows.Foundation.Collections
python-dotenv
jinja2